requirements.txt
nest_asyncio
aiohttp
aiohttp_session
aiohttp_security
numpy
pandas==1.4.3
scikit-learn
openai
checklist
shap==0.41.0
transformers
torch==1.12.0
spacy
profanity
Pillow
appdirs
diskcache
torchvision==0.13.0
clip-anytorch==2.4.0
clip-retrieval==2.34.2
google_images_search==1.4.3